# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Keuangan Simpadu API",
        "description": "Dokumentasi API Keuangan Simpadu",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tagihan": {
            "get": {
                "tags": [
                    "Tagihan"
                ],
                "summary": "Tampilkan semua tagihan",
                "operationId": "ea77da2ae8a2d401d97278ec9e99fe54",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id_tagihan": {
                                                "type": "integer"
                                            },
                                            "nim": {
                                                "type": "string"
                                            },
                                            "nama_tagihan": {
                                                "type": "string"
                                            },
                                            "id_thn_ak": {
                                                "type": "string"
                                            },
                                            "id_kategori_ukt": {
                                                "type": "integer"
                                            },
                                            "status_tagihan": {
                                                "type": "integer"
                                            },
                                            "tgl_terbit": {
                                                "type": "string",
                                                "format": "date"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tagihan"
                ],
                "summary": "Tambah tagihan baru",
                "operationId": "53a9865fe29f08fc1deb85893d621b79",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nim",
                                    "nama_tagihan",
                                    "id_thn_ak",
                                    "id_kategori_ukt",
                                    "tgl_terbit"
                                ],
                                "properties": {
                                    "nim": {
                                        "type": "string",
                                        "example": "1234567890123456"
                                    },
                                    "nama_tagihan": {
                                        "type": "string",
                                        "example": "UKT Semester Genap 2025"
                                    },
                                    "id_thn_ak": {
                                        "type": "string",
                                        "example": "20245"
                                    },
                                    "id_kategori_ukt": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "status_tagihan": {
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "tgl_terbit": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tagihan berhasil ditambahkan"
                    }
                }
            }
        },
        "/api/tagihan/{id}": {
            "get": {
                "tags": [
                    "Tagihan"
                ],
                "summary": "Detail tagihan",
                "operationId": "108530daf74cf26fcc8752ecba76a9ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id_tagihan": {
                                            "type": "integer"
                                        },
                                        "nim": {
                                            "type": "string"
                                        },
                                        "nama_tagihan": {
                                            "type": "string"
                                        },
                                        "kategori_ukt": {
                                            "type": "string"
                                        },
                                        "nominal": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Tagihan"
                ],
                "summary": "Update tagihan",
                "operationId": "bf2440a727cd689fa44e7c0718e22b40",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nim": {
                                        "type": "string"
                                    },
                                    "nama_tagihan": {
                                        "type": "string"
                                    },
                                    "id_thn_ak": {
                                        "type": "string"
                                    },
                                    "id_kategori_ukt": {
                                        "type": "integer"
                                    },
                                    "status_tagihan": {
                                        "type": "integer"
                                    },
                                    "tgl_terbit": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tagihan berhasil diupdate"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tagihan"
                ],
                "summary": "Hapus tagihan",
                "operationId": "0daf2ccb7d9a754152506c4162dafdb1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tagihan berhasil dihapus"
                    }
                }
            }
        },
        "/api/tagihan/{id}/nominal-akhir": {
            "get": {
                "tags": [
                    "Tagihan"
                ],
                "summary": "Hitung nominal akhir tagihan setelah potongan keringanan",
                "operationId": "abe0999ce5efcb149ac29f28c6914fb1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nominal akhir tagihan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id_tagihan": {
                                            "type": "integer"
                                        },
                                        "nim": {
                                            "type": "string"
                                        },
                                        "nama_tagihan": {
                                            "type": "string"
                                        },
                                        "kategori_ukt": {
                                            "type": "string"
                                        },
                                        "nominal_ukt": {
                                            "type": "integer"
                                        },
                                        "total_potongan": {
                                            "type": "integer"
                                        },
                                        "nominal_akhir": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}